Refactor Data backends into Link
This commit adds the `Link' class, which can wrap any array type that
in some way or form can be expose its data to ASTRA as either a numpy
array or GPULink object.

As before, there is a numpy link (which is enabled by default) and a
torch backend which can be enabled optionally.