Add VolumeVectorGeometry
The goal is to replace OrientedBox.

OrientedBox has two deficiences:
- the size and coordinate frame have overlapping use cases, leading to
  weird properties like rel_size and abs_size
- It cannot be converted to a VolumeGeometry in any way since it has
  no shape.

The VolumeVectorGeometry has these nice properties:
- It restores the symmetry between volume geometry on one side and
  parallel and cone geometries on the other.
- ts.operator could in principle accept VolumeVectorGeometries and
  transform the projection geometry into its coordinate frame.

So the plan is to:
- Add support for displaying VolumeVectorGeometries
- Remove OrientedBox completely
- Add support for VolumeVectorGeometries in ts.operator

At that point, ts.data would be the only object that does not support
VolumeVector geometries (since ASTRA does not support arbitrarily
oriented volumes). This is something to consider.