Unify VolumeGeometry and VolumeVectorGeometry
Since a VolumeGeometry is effectively a VolumeVectorGeometry in a
single axis-aligned orientation, VolumeGeometry now wraps an inner
VolumeVectorGeometry ("has-a" rather than "is-a" relationship).

The two classes now both implement the following properties:
- pos, w, v, u
- num_steps
- shape
- size, sizes
- voxel_size, voxel_sizes
- voxel_volume, voxel_volumes
- corners

Most of the computation is performed in VolumeVectorGeometry.

In addition, VolumeGeometry implements `extent`, which is a popular
way to parametrize geometries in ASTRA and ODL.