tests: Fix doctests in the presence of ODL
ODL adjusts NumPy's print options to make its own tests more
consistent. This is annoying when using doctests in tomosipo, as the
printed output may be different when ODL is installed in the same
environment.

This commit adds the UniformPrintingTestCase class that can be
sub-classed and makes sure that the print options are consistent.