doc: Fix mock imports
The PyTorch and CuPy linking facilities are auto-imported by Sphinx
autodoc. This leads to problems when the linking facilities are actually
used in running code.

Now, we first try to import a dependency before letting autodoc mock it.
Secondly, we detect if the cupy or torch modules are mocked and do not
add their linking facilities.