Make external array support automatic
Previously, to add support for torch or cupy arrays, you had to execute
the following

import tomosipo.torch_support
import tomsipo.cupy

This was done to prevent surprises when suddenly torch or cupy arrays
did or did not work because the parent packages were not installed.

This is a bit stupid, however, because if the torch package is not
installed, how can the user create a torch array? Their script would
probably crash around the `import torch` statement.

Therefore, the support for linking of foreign arrays has been made
automatic. If the parent package is available, linking to the array type
is supported.